requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
click==8.3.1
colorama==0.4.6
fastapi==0.123.4
greenlet==3.2.4
grpcio==1.76.0
grpcio-tools==1.76.0
h11==0.16.0
idna==3.11
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.23.1
protobuf==6.33.1
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dotenv==1.2.1
setuptools==80.9.0
SQLAlchemy==2.0.44
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
